requirements.txt
google-cloud-texttospeech
openai
pyttsx3
python-dotenv
pytz
requests
sounddevice
tiktoken
vosk
websockets
aider-install